refactor: move stake maturity
