+++ frontend/src/lib/api/sns.api.ts
@@ -405,30 +405,3 @@ export const stakeNeuron = async ({
   );
   return newNeuronId;
 };
-
-export const stakeMaturity = async ({
-  percentageToStake,
-  rootCanisterId,
-  identity,
-  neuronId,
-}: {
-  percentageToStake: number;
-  rootCanisterId: Principal;
-  identity: Identity;
-  neuronId: SnsNeuronId;
-}): Promise<void> => {
-  logWithTimestamp(`Stake maturity: call...`);
-
-  const { stakeMaturity: stakeMaturityApi } = await wrapper({
-    identity,
-    rootCanisterId: rootCanisterId.toText(),
-    certified: true,
-  });
-
-  await stakeMaturityApi({
-    neuronId,
-    percentageToStake,
-  });
-
-  logWithTimestamp(`Stake maturity: complete`);
-};
